requirements.txt
APScheduler==3.6.3
cachelib==0.1.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==7.1.2
dnspython==2.2.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
pymongo==4.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.27.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.8
Werkzeug==1.0.1
